Implement a label set abstraction
We're going to need to make the labels completely configurable in
order to follow the Probot best practices.

Initially, we're operating with two sets of labels:
roles (maintainer and author), and escalation (by level, with increasing
urgency).

I'm not convinced that this is the design we need, but it should be a
reasonable starting point to let us consolidate some of the magic
strings in the app, and then subsequently make the labels configurable.