Don't try to create labels that already exist
When we get an installation event, we create all the labels (roles and escalation) that the app
needs. However, if we try to create a label that already exists we get a 422.
The unnecessary call also counts against our rate limit, which is
another good reason not to make the call in the first place.